+++ src/features/providers/formatter/rangeFormat.ts
@@ -14,7 +14,17 @@ export class RangeFormattingProvider implements vscode.DocumentRangeFormattingEd
     range: vscode.Range,
     options: vscode.FormattingOptions,
     token: vscode.CancellationToken,
+  ): Promise<vscode.TextEdit[]> {
+    const textEdits = await FormatSelectionProvider.provideTextEdits(document, range);
 
+    return textEdits;
+  }
+}
+
+export class FormatSelectionProvider {
+  static async provideTextEdits(
+    document: vscode.TextDocument,
+    range: vscode.Range,
   ): Promise<vscode.TextEdit[]> {
     const filePath = Utilities.normalizePath(document.fileName);
     const workspaceFolder = vscode.workspace.workspaceFolders ? vscode.workspace.workspaceFolders[0].uri.fsPath : undefined;
@@ -22,12 +32,18 @@ export class RangeFormattingProvider implements vscode.DocumentRangeFormattingEd
     const workingDirectory = Configuration.workingDirectory(rootPath);
     const textEdits: vscode.TextEdit[] = [];
 
+    const endCharacter = document.lineAt(range.end.line).range.end.character;
+    const lineRange = new vscode.Range(
+      new vscode.Position(range.start.line, 0),
+      new vscode.Position(range.end.line, endCharacter),
+    );
+
     if (workingDirectory?.includes("${")) {
       return [];
     }
 
     Utilities.appendHyphenatedLine();
-    Utilities.outputChannel.appendLine(`Range (Lines ${range.start.line} to ${range.end.line}) Format triggered for ${filePath}`);
+    Utilities.outputChannel.appendLine(`Range (Lines ${lineRange.start.line} to ${lineRange.end.line}) Format triggered for ${filePath}`);
 
     if (!Configuration.formatEnabled()) {
       const message = "Format not enabled in the settings. Skipping Format.";
@@ -53,7 +69,7 @@ export class RangeFormattingProvider implements vscode.DocumentRangeFormattingEd
 
     const commandOptions: CommandOptions = {
       filePath: filePath,
-      fileContents: document.getText(range),
+      fileContents: document.getText(lineRange),
     };
 
     const result = await SQLFluff.run(
@@ -67,13 +83,18 @@ export class RangeFormattingProvider implements vscode.DocumentRangeFormattingEd
       throw new Error("Command failed to execute, check logs for details");
     }
 
-    const lines = FormatHelper.parseLines(result.lines);
+    let lines = FormatHelper.parseLines(result.lines);
+
+    const leadingWhitespace = document.lineAt(range.start.line).firstNonWhitespaceCharacterIndex;
+    lines = lines ? FormatHelper.addLeadingWhitespace(lines, document.languageId, leadingWhitespace) : undefined;
+
     if (lines === undefined) return [];
 
     if (lines.length > 1 || lines[0] !== "") {
+      const eol = document.eol;
       textEdits.push(vscode.TextEdit.replace(
-        range,
-        lines.join("\n"),
+        lineRange,
+        eol === vscode.EndOfLine.LF ? lines.join("\n") : lines.join("\r\n"),
       ));
     }